requirements.txt
fastapi
uvicorn
pdfminer.six
requests